requirements.txt
newspaper3k
scikit-learn
lxml
lxml_html_clean
Flask
SQLAlchemy
Flask-SQLAlchemy
flask-cors
requests